pyproject.toml
[tool.poetry]
name = "i-python"
version = "0.1.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uuid = "1.30"
python-socketio = "~5.7.1"
aiohttp = "^3.11.16"
loguru = "0.5.3"

[tool.poetry.group.front.dependencies]
pyqt5 = "^5.15.11"
aiohttp = "^3.13.2"

[tool.poetry.group.dev.dependencies]
mypy = "1.6.1"
flake8 = "^6.0.0"
dlint = "^0.14.1"
flake8-bugbear = "^23.6.5"
flake8-pie = "^0.16.0"
flake8-simplify = "^0.20.0"
flake8-warnings = "^0.4.0"
flake8-comprehensions = "^3.13.0"
flake8-eradicate = "^1.5.0"
black = "^23.7.0"
isort = "^5.12.0"


[tool.poetry.group.back.dependencies]
fastapi = "0.115.12"
uvicorn = "^0.34.0"
zmq = "0.0.0"
ipykernel="6.4.2"
python-multipart = "0.0.5"
pydantic = "1.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
